Refactor duplicated code
